+++ common/src/test/java/com/github/khanshoaib3/minecraft_access/test_utils/MockInterval.java
@@ -3,7 +3,8 @@ package com.github.khanshoaib3.minecraft_access.test_utils;
 import com.github.khanshoaib3.minecraft_access.utils.condition.Interval;
 
 public class MockInterval extends Interval {
-    boolean ready = false;
+    public static final MockInterval ALWAYS_READY = new MockInterval(0, 0);
+    boolean ready = true;
 
     public MockInterval(long lastRunTime, long delayInNanoTime) {
         super(lastRunTime, delayInNanoTime);