+++ doc/CHANGELOG.md
@@ -19,6 +19,11 @@ Release v1.2.2 (2023-09)
 ### Refactoring, Documentation and Chores
 
 * Add `CHANGELOG.md` for persistent change log and `PULL_REQUEST_TEMPLATE.md` for reminding PR submitters.
+* Move `PlayerPositionUtils`, `PositionUtils` into `utils.position` subpackage.
+* Move `MouseUtils`, `OSUtils`, `KeyUtils` into `utils.system` subpackage.
+* Add several reusable junit `*Extension` for testing.
+* Add `Orientation` enum for representing directions, refactor related code to apply it.
+* Add `IntervalKeystroke` for interval executed keystrokes.
 
 Release v1.2.1 (2023-08-31)
 ---------------------------